#4605 auto-creation of room users on login is now triggered by a dedicated subscriber class ("AutoRoomMembershipSubscriber") and not from within "LoginSubscriber"
this will ease writing subscriber tests or when priorising subscribers
