+++ src/EventSubscriber/LoginSubscriber.php
@@ -3,9 +3,6 @@
 namespace App\EventSubscriber;
 
 use App\Entity\Account;
-use App\Entity\AuthSource;
-use App\Entity\Portal;
-use App\Facade\UserCreatorFacade;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpKernel\Event\RequestEvent;
@@ -13,8 +10,6 @@ use Symfony\Component\HttpKernel\HttpKernelInterface;
 use Symfony\Component\HttpKernel\KernelEvents;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Component\Security\Core\Security;
-use Symfony\Component\Security\Http\Event\InteractiveLoginEvent;
-use Symfony\Component\Security\Http\SecurityEvents;
 
 class LoginSubscriber implements EventSubscriberInterface
 {
@@ -28,26 +23,18 @@ class LoginSubscriber implements EventSubscriberInterface
      */
     private UrlGeneratorInterface $urlGenerator;
 
-    /**
-     * @var UserCreatorFacade
-     */
-    private UserCreatorFacade $userCreator;
-
     public function __construct(
         Security $security,
-        UrlGeneratorInterface $urlGenerator,
-        UserCreatorFacade $userCreator
+        UrlGeneratorInterface $urlGenerator
     ) {
         $this->security = $security;
         $this->urlGenerator = $urlGenerator;
-        $this->userCreator = $userCreator;
     }
 
     public static function getSubscribedEvents()
     {
         return [
             KernelEvents::REQUEST => 'onKernelRequest',
-            SecurityEvents::INTERACTIVE_LOGIN => 'onSecurityInteractiveLogin',
         ];
     }
 
@@ -78,48 +65,5 @@ class LoginSubscriber implements EventSubscriberInterface
         if ($account->hasLegacyPassword()) {
             $event->setResponse(new RedirectResponse($this->urlGenerator->generate('app_migration_password')));
         }
-    }
-
-    public function onSecurityInteractiveLogin(InteractiveLoginEvent $event)
-    {
-        /** @var Account $account */
-        $account = $event->getAuthenticationToken()->getUser();
-        if (!$account instanceof Account) {
-            return;
-        }
-
-        /** @var AuthSource $uthSource */
-        $authSource = $account->getAuthSource();
-        if (!$authSource) {
-            return;
-        }
-
-        // check if auto-creating room memberships is enabled in the portal configuration
-        /** @var Portal $portal */
-        $portal = $authSource->getPortal();
-        if (!$portal || !$portal->getAuthMembershipEnabled()) {
-            return;
-        }
-
-        // extract any room identifiers from the request using the parameter key defined in the portal configuration
-        $membershipIdentifiersKey = $portal->getAuthMembershipIdentifier();
-        if (empty($membershipIdentifiersKey)) {
-            return;
-        }
-
-        $request = $event->getRequest();
-        $paramBag = $request->request;
-        $membershipIdentifierString = $paramBag->get($membershipIdentifiersKey);
-        if (empty($membershipIdentifierString)) {
-            return;
-        }
-
-        $membershipIdentifiers = explode(',', trim($membershipIdentifierString));
-        if (empty($membershipIdentifiers)) {
-            return;
-        }
-
-        // for the given account, create users in the specified rooms
-        $this->userCreator->addUserToRoomsWithSlugs($account, $membershipIdentifiers);
     }
 }
